Split test_vocab_examples.py into 5 focused test files
Split the 913-line monolithic TestVocabExamples class into 5 files
grouped by domain area:

- test_vocab_utils.py (2 tests): json2md/obj_to_file utilities
- test_vocab_actor_report_examples.py (10 tests): actors + reports
- test_vocab_participant_examples.py (14 tests): participants, status,
  invites, actor recommendations
- test_vocab_case_examples.py (22 tests): case lifecycle, notes,
  ownership transfer, case status — split into 6 inner test classes
- test_vocab_embargo_examples.py (9 tests): embargo lifecycle

All 57 tests pass. Each new file is ≤350 lines. The inline import for
_RecommendActorActivity was promoted to module-level.

Closes #496
